// web/src/lib/types/ApiDataField.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ApiDataField = { 
    category: string, 
    path: string, 
    semantic_type: string, 
    description: string, 
    icon: string | null,
    unit: string | null,
    privacy: string | null,
    confidence: number | null,
    temporal: string | null,
};